pyproject.toml
[tool.poetry]
name = "matrix-asgi"
version = "0.1.0"
description = "ASGI Server for the Matrix protocol"
authors = ["Guilhem Saurel <guilhem.saurel@laas.fr>"]
license = "BSD-2-Clause"

[tool.poetry.dependencies]
python = "^3.9"
channels = "^3.0.5"
matrix-nio = "^0.19.0"
Markdown = "^3.4.1"
Django = ">=4.1"

[tool.poetry.dev-dependencies]
channels-redis = "^3.4.1"
coverage = "^6.4.2"
pydocstyle = {extras = ["toml"], version = "^6.1.1"}
httpx = "^0.23.0"
safety = "^2.1.1"
flake8 = "^5.0.4"
black = "^22.6.0"
pyupgrade = "^2.37.3"

[tool.poetry.scripts]
matrix-asgi = "matrix_asgi.__main__:main"

[tool.pydocstyle]
match_dir = "(?=[^.])(?!migrations).*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
